requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Compress==1.14
mysql-connector-python==8.2.0
pandas==2.1.4
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.0
tqdm==4.65.0
